requirements-dev.txt
argparse
gitpython
loguru
pillow
pynacl
requests
toml
